Fixed lots of things and started others
- Fixed compilation errors
- Started new motor calibration
- Fixed and improved IK & Control Loop

+++ Firmware/include/drivers/MotorDriver.hpp
@@ -5,7 +5,7 @@
 namespace MotorDriver
 {
     using Channel = uint8_t;
-    using Value = uint16_t;
+    using Value = int;
     constexpr Channel CHANNEL_COUNT = 16;
 
     /**
